\name{[-methods}
\docType{methods}
\alias{[-methods}
\alias{[,FlexibleLagged,ANY-method}
\alias{[,FlexibleLagged,missing-method}
\alias{[,Lagged1d,numeric-method}
\alias{[,Lagged2d,numeric-method}
\alias{[,Lagged3d,numeric-method}
\alias{[,Lagged,missing-method}
\alias{[,ANY,ANY-method}
\alias{[,nonStructure,ANY-method}
\alias{[,slMatrix,ANY-method}
\title{Methods for subsetting defined in package 'lagged'}
\description{Methods for subsetting  defined in package 'lagged'.}
\section{Methods}{
  The following methods for \code{"["} are currently defined in package
  "lagged": 
  \describe{

    \item{\code{signature(x = "FlexibleLagged", i = "missing")}}{
      %%  ~~describe this method here~~
    }

    \item{\code{signature(x = "FlexibleLagged", i = "ANY")}}{
      %%  ~~describe this method here~~
    }

    \item{\code{signature(x = "Lagged1d", i = "numeric")}}{
      %%  ~~describe this method here~~
    }
    
    \item{\code{signature(x = "Lagged2d", i = "numeric")}}{
      %%  ~~describe this method here~~
    }
    
    \item{\code{signature(x = "Lagged3d", i = "numeric")}}{
      %%  ~~describe this method here~~
    }

    \item{\code{signature(x = "Lagged", i = "missing")}}{
      %%  ~~describe this method here~~
    }
    \item{\code{signature(x = "slMatrix", i = "ANY")}}{{
	Describe any discrepancy with the Lagged methods.
      }
}}}
\keyword{methods}
